Made the player appear onto the screen once the player clicks start. Made some changes to the player class as a result to accomodate this change.

+++ classes/Player.h
@@ -10,7 +10,7 @@ private:
 	float laneOneY;
 	float laneTwoY;
 	float laneThreeY;
-	int fixedX;
+	float fixedX;
 
 	// Player Data
 	cocos2d::Vec2		pos;
